+++ Models/Server/PhysicsItemDTO.cs
@@ -9,8 +9,8 @@ namespace MapEditor.Models.Server
         [JsonProperty("roomId"), CustomProperty("Room Id", true)] public int RoomId { get; set; }
         [JsonProperty("inGamePhysicsItemId")] public int InGamePhysicsItemId { get; set; }
         [JsonProperty("name"), CustomProperty] public required string Name { get; set; } = "Physics Item";
-        [JsonProperty("roomOffsetX"), CustomProperty("X")] public float RoomOffsetX { get; set; }
-        [JsonProperty("roomOffsetY"), CustomProperty("Y")] public float RoomOffsetY { get; set; }
+        [JsonProperty("roomOffsetX"), CustomProperty("X")] public int RoomOffsetX { get; set; }
+        [JsonProperty("roomOffsetY"), CustomProperty("Y")] public int RoomOffsetY { get; set; }
         [JsonProperty("simulatePhysics"), CustomProperty("Simulate Physics")] public bool SimulatePhysics { get; set; }
         [JsonProperty("isRotationRandom"), CustomProperty("Rotation Random")] public bool IsRotationRandom { get; set; }
         [JsonProperty("zAngleRotation"), CustomProperty("Rotation", showIfProperty: "IsRotationRandom", showIfValue: false)] public float ZAngleRotation { get; set; }
